apps/ollama-pod/pyproject.toml
[project]
name = "ollama-pod"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "typer",
    "runpod",
    "httpx",
    "python-dotenv",
    "rich",
]

[project.scripts]
ollama-pod = "ollama_pod.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ollama_pod"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
